Move compiler plugin's code around
